fix: link to deletedfile
