+++ src/components/Layout.tsx
@@ -32,9 +32,9 @@ const useStyles = makeStyles((/* theme: Theme */) =>
       position: 'relative' as 'relative',
       background: 'white',
       top: '100px',
-      maxWidth: '1024px',
+      maxWidth: '1200px',
       margin: '0 auto',
-      padding: '50px 0',
+      padding: '50px 10px',
       zIndex: 2,
       marginBottom: '100px',
     },
