--- src/models/evaluation.ts
@@ -1,10 +1,0 @@
-export interface EvaluationModel {
-  title: string;
-  tasks: EvaluationTaskModel[];
-}
-
-export interface EvaluationTaskModel {
-  title: string;
-  score: number;
-  done: boolean;
-}
