--- src/models/worklogModel.ts
@@ -1,9 +1,0 @@
-export interface TaskModel {
-  title: string;
-  hours: number;
-}
-
-export interface WorklogModel {
-  name: string;
-  worklog: TaskModel[];
-}
